Add files via upload
In this notebook we used a simple linear regression to predict the price a group of houses in Iowa.

Quantitative and qualitative variables were used for modeling, the qualitative ones were converted into dummies for being used into our model. The strenght of linear correlation with the target variable was used to choose the quantitative variables used. In the case of the qualitative ones, an ANNOVA analysis was used.

We detected the outliers using a univariate analysis with the z-core value as indicator.Functions as logarithms and square root were used during the process to get normality and homoscedasticity of the variables.